move MQ listener to main
